Fixing bug #915
Wildcard ignores cause ag to discard all result files.  This commit
addresses this issue by first checking if a wildcard '*' gitignore is
present then performs a negative check against all files that would be
disqualified by the wildcard ignore with any negative-ignores
(represented by entries starting with '!' in a .gitignore file).  If a
negative match is found, the previous match against '*' is skipped.